feat: 지부 삭제 request body 변경

+++ src/main/java/com/umc/networkingService/domain/branch/service/BranchUniversityService.java
@@ -27,14 +27,14 @@ public class BranchUniversityService {
 
     //지부, 대학 연결하기
     @Transactional
-    public void connectBranchUniversity(UUID branchId, List<UUID> request) {
-
-        request.forEach(universityId -> {
+    public void connectBranchUniversity(UUID branchId, List<UUID> universityIds) {
+        //유효한 지부가 값으로 들어옴
+        universityIds.forEach(universityId -> {
             if(!isUniversityValid(universityId)){ //유효한 대학인지 확인
                 throw new BranchUniversityHandler(ErrorCode.UNIVERSITY_NOT_FOUND);
             }
             if (isBranchUniversityExist(branchId, universityId)) { //지부-대학 연결 여부 확인
-                throw new BranchUniversityHandler(ErrorCode.BRANCH_UNIVERSITY_ALREADY_EXIST);
+                throw new BranchUniversityHandler(ErrorCode.BRANCH_UNIVERSITY_ALREADY_EXIST); //이미 연결됨
             }
             branchUniversityRepository.save
                     (
@@ -48,12 +48,17 @@ public class BranchUniversityService {
 
     //지부, 대학 연결 끊기
     @Transactional
-    public void disconnectBranchUniversity(UUID branchId, UUID universityId) {
-        //유효한 대학, 지부가 값으로 들어옴
-        if (!isBranchUniversityExist(branchId, universityId)) { //지부-대학 연결 여부 확인
-            throw new BranchUniversityHandler(ErrorCode.BRANCH_UNIVERSITY_NOT_FOUND);
-        }
-        branchUniversityRepository.deleteByBranchIdAndUniversityId(branchId, universityId);
+    public void disconnectBranchUniversity(UUID branchId, List<UUID> universityIds) {
+        //지부가 값으로 들어옴
+        universityIds.forEach(universityId -> {
+            if(!isUniversityValid(universityId)){ //유효한 대학인지 확인
+                throw new BranchUniversityHandler(ErrorCode.UNIVERSITY_NOT_FOUND);
+            }
+            if (!isBranchUniversityExist(branchId, universityId)) { //지부-대학 연결 여부 확인
+                throw new BranchUniversityHandler(ErrorCode.BRANCH_UNIVERSITY_NOT_FOUND); //연결되지 않음
+            }
+            branchUniversityRepository.deleteByBranchIdAndUniversityId(branchId, universityId);
+        });
     }
 
     //지부-대학 연결 여부 확인
